Added Watched/Watching, etc. functionality, but it doesnt work as I'd like it to. but it works

+++ anime-reviews/src/main/java/anime_reviews/dao/AnimeReviewDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import anime_reviews.entity.AnimeReview;
 
-public interface AnimeReviewDAO extends JpaRepository<AnimeReview, Integer> {
+public interface AnimeReviewDAO extends JpaRepository<AnimeReview, Long> {
     
 }
