+++ anime-reviews/src/main/java/anime_reviews/dao/UserDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import anime_reviews.entity.Users;
 
-public interface UserDAO extends JpaRepository<Users, Integer>{
+public interface UserDAO extends JpaRepository<Users, Long>{
     
 }
